Refaktorisanje frontend aplikacije
