+++ ktsnwt-client/src/app/config/router-paths.ts
@@ -1,18 +1,30 @@
 export const HOME_PATH = '';
+
+// User
 export const LOGIN_PATH = 'login';
 export const REGISTRATION_PATH = 'register';
 export const VERIFY_ACCOUNT_PATH = 'verify';
 export const USER_SETTINGS_PATH = 'settings';
+
+// Locations
 export const LOCATIONS_PATH = 'locations';
-export const EVENTS_PATH = 'events';
 export const ADD_LOCATION_PATH = 'add-location';
 export const EDIT_LOCATION_PATH = 'edit-location/:id';
-export const MY_RESERVATIONS_PATH = 'my-reservations';
+
+// Events
+export const EVENTS_PATH = 'events';
 export const ADD_EVENT_PATH = 'add-event';
 export const SHOW_EVENT_DETAILED_ID = 'show-event/:id';
 export const SHOW_EVENT_DETAILED = 'show-event/';
+
+// Halls
 export const HALL_SETTINGS = 'hall-settings/:id';
-export const NOT_FOUND = 'not-found';
-export const NOT_AUTHORIZED = 'not-authorized';
+
+// Reservations
+export const MY_RESERVATIONS_PATH = 'my-reservations';
 export const RESERVE_TICKET = 'reserve/';
 export const RESERVE_TICKET_ID = 'reserve/:eventId/:eventDayId';
+
+// Error pages
+export const NOT_FOUND = 'not-found';
+export const NOT_AUTHORIZED = 'not-authorized';
