+++ ktsnwt-client/src/app/guards/auth.guard.ts
@@ -1,20 +1,19 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree, Router } from '@angular/router';
 import { Observable } from 'rxjs';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../services/auth.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class AuthLoginGuard implements CanActivate {
+export class AuthGuard implements CanActivate {
 
-  constructor(private authService: AuthService,
-              private router: Router) {}
+  constructor(private authService: AuthService, private router: Router) { }
 
   canActivate(
     next: ActivatedRouteSnapshot,
     state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-      if (!this.authService.isUserLoggedIn()) {
+      if (!this.authService.checkForAdmin()) {
         this.router.navigate(['/not-authorized']);
         return false;
       }
