+++ ktsnwt-client/src/app/guards/edit-location.guard.ts
