+++ ktsnwt-client/src/app/guards/show-event.guard.ts
