+++ ktsnwt-client/src/app/models/request/event-info.ts
