+++ ktsnwt-client/src/app/models/request/login-info.ts
