+++ ktsnwt-client/src/app/models/request/user-registration-data.ts
@@ -1,4 +1,4 @@
-export interface UserInfo {
+export interface UserRegistrationData {
     username: string;
     password: string;
     repeatPassword: string;
