+++ ktsnwt-client/src/app/models/response/user.ts
