+++ ktsnwt-client/src/app/services/auth.service.ts
@@ -1,7 +1,7 @@
-import { UserInfo } from './../models/user-info';
+import { UserRegistrationData } from '../models/request/user-registration-data';
 import { API_LOGIN, API_REGISTER_USER, API_VERIFY_ACCOUNT } from './../config/api-paths';
 import { Observable } from 'rxjs';
-import { LoginInfo } from '../models/login-info';
+import { LoginInfo } from '../models/request/login-info';
 import { USER_ID_KEY, USER_ROLE_KEY, USERNAME_KEY, USER_TOKEN_KEY } from './../config/local-storage-keys';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
@@ -33,7 +33,7 @@ export class AuthService {
     localStorage.removeItem(USER_TOKEN_KEY);
   }
 
-  addNewUser(userInfo: UserInfo): Observable<any> {
+  addNewUser(userInfo: UserRegistrationData): Observable<any> {
     return this.http.post(API_REGISTER_USER, userInfo);
   }
