--- ktsnwt-client/src/app/services/auth/auth-login.guard.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, async, inject } from '@angular/core/testing';
-
-import { AuthLoginGuard } from './auth-login.guard';
-
-describe('AuthLoginGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AuthLoginGuard]
-    });
-  });
-
-  it('should ...', inject([AuthLoginGuard], (guard: AuthLoginGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
