--- ktsnwt-client/src/app/services/auth/edit-location.guard.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, async, inject } from '@angular/core/testing';
-
-import { EditLocationGuard } from './edit-location.guard';
-
-describe('EditLocationGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [EditLocationGuard]
-    });
-  });
-
-  it('should ...', inject([EditLocationGuard], (guard: EditLocationGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
