--- ktsnwt-client/src/app/services/auth/show-event.guard.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, async, inject } from '@angular/core/testing';
-
-import { ShowEventGuard } from './show-event.guard';
-
-describe('ShowEventGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ShowEventGuard]
-    });
-  });
-
-  it('should ...', inject([ShowEventGuard], (guard: ShowEventGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
