+++ ktsnwt-client/src/app/services/location.service.ts
@@ -27,14 +27,14 @@ export class LocationService {
 
   addLocation(locationData): Observable<any> {
     return this.http.post(API_LOCATION_ADD, locationData).pipe(
-      tap(data => console.log(data)),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
 
   getLocationById(locationId): Observable<any> {
     return this.http.get(API_LOCATION_GET_ID + '/' + locationId).pipe(
-      tap(data => console.log(data)),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
@@ -49,14 +49,14 @@ export class LocationService {
 
   editLocation(locationData): Observable<any> {
     return this.http.put(API_LOCATION, locationData).pipe(
-      tap(data => console.log(JSON.stringify(data))),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
 
   changeAddress(addressData, locationId): Observable<any> {
     return this.http.put(API_LOCATION_ADDRESS + '/' + locationId, addressData).pipe(
-      tap(data => console.log(JSON.stringify(data))),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
@@ -65,28 +65,28 @@ export class LocationService {
 
   getHallById(hallId): Observable<any> {
     return this.http.get(API_HALL_GET + '/' + hallId).pipe(
-      tap(data => console.log(data)),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
 
   addHall(hallData): Observable<any> {
     return this.http.post(API_HALL_GET, hallData).pipe(
-      tap(data => console.log(data)),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
 
   addSector(sectorData): Observable<any> {
     return this.http.post(API_SECTOR, sectorData).pipe(
-      tap(data => console.log(data)),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
 
   modifySector(sectorData): Observable<any> {
     return this.http.put(API_SECTOR, sectorData).pipe(
-      tap(data => console.log(data)),
+      tap(data => {}),
       catchError(this.handleError)
     );
   }
